refactor

+++ src/helpers/deleteImageURL.ts
@@ -1,10 +1,14 @@
+import { config } from "firebase-functions";
 import fetch from "node-fetch";
-import { getConfig } from "../utils/getConfig";
 
 export const deleteImageURL = async (bucketName: string, filePath: string) => {
-  const config = getConfig();
+  const appConfig = config().app;
 
-  await fetch(config.app.images, {
+  if (!appConfig || !appConfig.images) {
+    return null;
+  }
+
+  await fetch(appConfig.images, {
     method: "DELETE",
     body: JSON.stringify({ bucketName, filePath }),
     headers: { "Content-Type": "application/json" }
